rename class to avoid merge conflict with other implementations
